binance-db/pyproject.toml
[project]
name = "binance-public-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "asyncio>=4.0.0",
    "duckdb>=1.3.2",
    "fire>=0.7.0",
    "pandas>=2.3.1",
    "pyarrow>=21.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]
